+++ include/queue/static_queue.h
@@ -5,19 +5,19 @@
 #include <stdbool.h>
 
 typedef void* static_queue_t;
-typedef uint8_t static_queue_element_t;
+typedef unsigned int static_queue_element_t;
 
 /**
  * Initialize a static queue of `n` elements
  *
  * @return positive integer if success, -1 if operation fails
  */
-int8_t static_queue_init(static_queue_t *squeue, uint8_t n);
+int static_queue_init(static_queue_t *squeue, unsigned int n);
 
 /**
  * Size
  */
-uint8_t static_queue_length(const static_queue_t squeue);
+unsigned int static_queue_length(const static_queue_t squeue);
 bool static_queue_is_empty(const static_queue_t squeue);
 bool static_queue_is_full(const static_queue_t squeue);
 
@@ -25,7 +25,7 @@ bool static_queue_is_full(const static_queue_t squeue);
  * Queue operations
  */
 static_queue_element_t static_queue_front(const static_queue_t squeue);
-int8_t static_queue_enqueue(static_queue_t squeue, static_queue_element_t elem);
+int static_queue_enqueue(static_queue_t squeue, static_queue_element_t elem);
 static_queue_element_t static_queue_dequeue(static_queue_t squeue);
 
 void static_queue_clear(static_queue_t squeue);
