+++ include/set/set.h
@@ -16,20 +16,20 @@ typedef enum { DEFAULT_SET, BIT_VECTOR_SET, LINKED_LIST_SET } set_type_t;
  * Initialize an empty set with a universe consisting of 0..n-1 integer elements.
  * @return positive integer or 0 if success, negative one if fails
  */
-int8_t set_init(set_t *set, int n, set_type_t type);
+int set_init(set_t *set, int n, set_type_t type);
 
 /**
  * Size
  */
-uint8_t set_length(const set_t set);
+unsigned int set_length(const set_t set);
 bool set_is_empty(const set_t set);
 
 /**
  * Membership, insertion and deletion operations on a set `set`.
  */
 bool set_member(set_t set, set_element_t e);
-int8_t set_insert(set_t set, set_element_t e);
-int8_t set_delete(set_t set, set_element_t e);
+int set_insert(set_t set, set_element_t e);
+int set_delete(set_t set, set_element_t e);
 void set_clear(set_t set);
 
 /**
@@ -38,9 +38,9 @@ void set_clear(set_t set);
  * not initialized, unexpected behavior might happen.
  * @return positive integer or 0 if success, negative one if fails
 */
-int8_t set_union(const set_t a, const set_t b, set_t *c);
-int8_t set_intersection(const set_t a, const set_t b, set_t *c);
-int8_t set_difference(const set_t a, const set_t b, set_t *c);
+int set_union(const set_t a, const set_t b, set_t *c);
+int set_intersection(const set_t a, const set_t b, set_t *c);
+int set_difference(const set_t a, const set_t b, set_t *c);
 
 /**
  * Destroy the set `set`. It must be called when the set won't be used
